requirements.txt
selenium
pytest
pytest-xdist
webdriver-manager
allure-pytest
python-dotenv
requests        # для API тестов
PyMySQL             # для подключения к MySQL
pytest-mysql        # фикстуры для MySQL (опционально, но удобно)